requirements.txt
# Core RAG components
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
qdrant-client>=1.6.4
sentence-transformers>=2.2.2
openai>=1.3.0
tiktoken>=0.5.1

# Database connectors
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
pymysql>=1.1.0
odoo-rpc-client>=0.9.0
odoorpc>=0.9.0

# Communication platforms connectors
mattermostdriver>=7.3.2
imaplib2>=3.6
exchangelib>=5.0.3
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Document processing
pypdf>=3.15.1
python-docx>=0.8.11
python-pptx>=0.6.21
markdown>=3.5.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
html2text>=2020.1.16
unstructured>=0.10.30
pytesseract>=0.3.10
pdfminer.six>=20221105

# Data formats
pydantic>=2.4.2
jsonlines>=3.1.0
pyyaml>=6.0.1
toml>=0.10.2
xmltodict>=0.13.0

# Graph and vector processing
networkx>=3.1
neo4j>=5.13.0
spacy>=3.7.1
scikit-learn>=1.3.2
numpy>=1.24.0
pandas>=2.1.2

# Web and API 
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.1
starlette>=0.27.0
jinja2>=3.1.2
requests>=2.31.0
aiohttp>=3.8.5

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
tqdm>=4.66.1
loguru>=0.7.2
click>=8.1.7
pytest>=7.4.3
pyvis>=0.3.2
matplotlib>=3.7.3

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
nbformat>=5.9.2